refactor: expose connection params [force release]
Linked to CU-86bxw199m
