+++ packages/datasource-mongo/src/index.ts
@@ -1,5 +1,5 @@
 import type { Introspection } from './introspection/types';
-import type { IntrospectorParams, MongoDatasourceParams } from './type';
+import type { ConnectionParams, IntrospectorParams, MongoDatasourceParams } from './types';
 import type { DataSourceFactory, Logger } from '@forestadmin/datasource-toolkit';
 import type { Connection } from 'mongoose';
 
@@ -12,6 +12,7 @@ import OdmBuilder from './odm-builder';
 export type { Introspection };
 export type { IntrospectorParams };
 export type { MongoDatasourceParams };
+export type { ConnectionParams };
 
 export async function introspect(options: IntrospectorParams): Promise<Introspection> {
   const { uri, connection: connectOptions, introspection: introspectionOptions } = options;
